new_blogV1.0/requirements.txt
gunicorn
gevent
flask
mysql.connector